+++ client/app/queue/pulacCerullo/PulacCerulloReminderModal.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 
-import Modal from '../components/Modal';
-import COPY from '../../COPY.json';
-import RadioField from '../components/RadioField';
+import Modal from '../../components/Modal';
+import COPY from '../../../COPY.json';
+import RadioField from '../../components/RadioField';
 import { CavcLinkInfo } from './CavcLinkInfo';
 
 const title = COPY.PULAC_CERULLO_REMINDER_MODAL_TITLE;
